refactor: groupBy to functions instead of helpers
