+++ lib/functions.js
@@ -1,6 +1,7 @@
 let logLevel = 3;
 
-export const getYYYYMMDDHHMMSS = (date = new Date()) => [
+export const getYYYYMMDDHHMMSS = (date = new Date()) =>
+  [
     date.getUTCFullYear(),
     (date.getUTCMonth() + 1).toString().padStart(2, '0'),
     date.getUTCDate().toString().padStart(2, '0'),
@@ -31,8 +32,23 @@ export const getFileName = (path) => path.replace(/^.*(\\|\/|:)/, '');
 
 export const log = (logLvl, message) => {
   if (logLvl >= logLevel) console.log(message);
-}
+};
 
 export const setLogLevel = (logLvl) => {
   logLevel = logLvl;
 };
+
+/**
+ * Slightly modified lodash groupBy
+ * https://github.com/lodash/lodash/blob/master/groupBy.js
+ */
+export const groupBy = (collection, iteratee) =>
+  collection.reduce((result, value, key) => {
+    key = iteratee(value);
+    if (Object.hasOwnProperty.call(result, key)) {
+      result[key].push(value);
+    } else {
+      result[key] = [value];
+    }
+    return result;
+  }, {});
