+++ lib/helpers.js
@@ -258,21 +258,6 @@ export const pathConfig = (options) => {
 export const readAsset = (assetPath) =>
   fs.readFileSync(join(dirname, './assets', assetPath)).toString();
 
-/**
- * Slightly modified lodash groupBy
- * https://github.com/lodash/lodash/blob/master/groupBy.js
- */
-export const groupBy = (collection, iteratee) =>
-  collection.reduce((result, value, key) => {
-    key = iteratee(value);
-    if (Object.hasOwnProperty.call(result, key)) {
-      result[key].push(value);
-    } else {
-      result[key] = [value];
-    }
-    return result;
-  }, {});
-
 // #region new method for sorting actions / NEEDS More TESTING
 /*
   actions.sort((a, b) => {
