+++ lib/models.js
@@ -3,7 +3,7 @@ import inquirer from 'inquirer';
 import hash from 'object-hash';
 
 import { parseAttributes, parseIndex } from './parser.js';
-import { groupBy } from './helpers.js';
+import { groupBy } from './functions.js';
 
 const { log } = console;
 
@@ -154,7 +154,6 @@ export const getActions = async (oldTables, newTables, upActions = null) => {
     possibleRenames,
     upActions,
   );
-
 
   Object.values(difference).forEach((df) => {
     const action = { tableName: df.path[0], depends: [df.path[0]], options: {} };
